refactor(node/state_delta): extract crypto + verify submodules from the god-file handler
`handlers/state_delta/mod.rs` had grown to 3326 lines mixing many
concerns. This is increment 1 of splitting it into focused submodules —
behavior-preserving, pure code movement (mod.rs: +4 wiring lines, -264).

- `crypto.rs` — `lookup_group_key_with_wait` (+ its key-lookup constants)
  and `decrypt_delta_actions`, with their unit tests. Leaf helpers with
  no external callers; re-imported into mod.rs so call sites are unchanged.
- `verify.rs` — `GroupIdCheck` (+ Debug) and
  `verify_position_group_id_matches_context`, the anti-bypass group_id
  check shared with `sync::manager` / `sync::delta_request`. Re-exported
  `pub(crate)` so the existing `crate::handlers::state_delta::{...}` import
  paths and the in-place `group_id_check_tests` continue to resolve.

No logic changes. Validated: node lib tests (274) and sync_sim (314) pass.

Remaining clusters for follow-up increments: event execution
(parse/execute/emit), governance buffering/replay, and delta-store setup.
